fix(ci): make sure to generateUsage help files before triggering build
this makes sure that when README.md is updated with latest version
number(s), there is also an up-to-date collection of usage help files
available. This is required, because even when using a dependency to
`generateUsage` task, the config phase of spotlessMarkdownApply has
already picked up the empty/non-existing files, even though they would
be available when the task is actually run...
so by generating the files before hand, we make sure they are available
in config phase of spotlessMarkdownApply.